tui: Move help display
Interactive mode now delegates flow parsing, command escapes, scoped menus,
and hunk routing outside the main dispatcher.

The help renderer still keeps static command text and color formatting
inside interactive.py, which blocks moving action dispatch without an
import cycle.

This commit addresses that by moving interactive help rendering into
tui.help_display and importing print_help from that module.

Subsequent commits will retarget help tests and guard the display owner.
